use double precision explicitly

+++ hoomd/hpmc/update.py
@@ -488,7 +488,7 @@ class RemoveDrift(Updater):
         super().__init__(trigger)
         self._param_dict.update({
             "reference_positions":
-                hoomd.data.typeconverter.NDArrayValidator(float, (None, 3))
+                hoomd.data.typeconverter.NDArrayValidator(np.float64, (None, 3))
         })
         self.reference_positions = reference_positions
